add: 오늘의 choice component
