+++ src/components/TodaysChoice/style.ts
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+
+export const TodaysChoice = styled.section`
+  margin-left: 10%;
+`;
+
+export const TodaysChoiceLayout = styled.div`
+  display: flex;
+  flex-direction: column;
+  row-gap: 5rem;
+`;
+
+export const Todays = styled.h1`
+  width: 13rem;
+  font-size: 2.2rem;
+  line-height: 4rem;
+  margin: 3.4rem 0;
+  border-bottom: 1px solid #000;
+`;
