+++ src/pages/PostDetail/style.ts
@@ -139,24 +139,6 @@ export const VoteButton = styled.button`
   }
 `;
 
-export const TodaysPosts = styled.section`
-  margin-left: 10%;
-`;
-
-export const TodaysPostsLayout = styled.div`
-  display: flex;
-  flex-direction: column;
-  row-gap: 5rem;
-`;
-
-export const Todays = styled.h1`
-  width: 13rem;
-  font-size: 2.2rem;
-  line-height: 4rem;
-  margin: 3.4rem 0;
-  border-bottom: 1px solid #000;
-`;
-
 export const CommentSection = styled.section`
   text-align: left;
   padding-top: 7rem;
